✅ differential testing between node and sage
This commit implements a differential test between the implementation in node
and the reference implementation written in sage. The test is responsible for
generating a random publickey on the secp256r1 curve and then running the
precomputation on both implementations. The output of both implementations is
then compared to ensure that they are the same. As the implementation in SAGE
is considered the reference implementation, it is critic this test never fails.
In addition to that, the main script has been cleaned up and the unused
scripts have been removed.