📦️ move from swc to parcel for building
Removed `swc` and `@swc/core` from the project and replaced it with
`parcel` for building the project. `@swc/jest` is still used for testing.
Configure parcel to build the project as an ES module. Create a new tsconfig
file as it was required. Take the opportunity to migrate .js script files to
.mjs files.